+++ src/index.ts
@@ -11,6 +11,7 @@ export * from "./components/alerts";
 export * from "./components/navbar";
 export * from "./components/disclosure";
 export * from "./components/charts";
+export * from "./components/marketing";
 export { Modal, StepModal } from "./components/Modal";
 export type { ModalProps, StepModalProps, ModalStep } from "./components/Modal";
 export { Loader } from "./components/Loader";